Move some things around for better structuring
