+++ app/src/main/java/com/g/tragosapp/application/AppConstants.kt
@@ -1,4 +1,4 @@
-package com.g.tragosapp.utils
+package com.g.tragosapp.application
 
 /**
  * Created by Gastón Saillén on 02 August 2020
@@ -9,5 +9,5 @@ object AppConstants {
     const val BASE_URL = "https://www.thecocktaildb.com/api/json/v1/1/"
 
     //ROOM
-    const val TABLE_NAME = "tabla_tragos"
+    const val DATABASE_NAME = "tabla_tragos"
 }
