+++ app/src/main/java/com/g/tragosapp/core/Resource.kt
@@ -1,4 +1,4 @@
-package com.g.tragosapp.vo
+package com.g.tragosapp.core
 
 sealed class Resource<out T> {
     class Loading<out T> : Resource<T>()
