+++ app/src/main/java/com/g/tragosapp/ui/MainViewModel.kt
@@ -4,9 +4,9 @@ import androidx.hilt.Assisted
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.*
 import com.g.tragosapp.application.ToastHelper
+import com.g.tragosapp.core.Resource
 import com.g.tragosapp.data.CocktailDataSource
 import com.g.tragosapp.data.model.Cocktail
-import com.g.tragosapp.vo.Resource
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
